+++ CHANGELOG.md
@@ -13,6 +13,8 @@
 
 - [CMS-1419] Added Wagtail registration view for SSO users, group info management UI, and updated user edit view to mark users as approved
 - [CMS-1421] Integrate with GOV.UK Notify to send SSO user approval notifications
+- [CI-104] - Added capital invest landing, region, sector and opportunity pages.
+
 
 ## [2019.05.09](https://github.com/uktrade/directory-cms/releases/tag/2019.05.09)
 [Full Changelog](https://github.com/uktrade/directory-cms/compare/2019.04.24_1...2019.05.09)